+++ app/src/main/java/com/codesoom/assignment/advice/BadRequestErrorAdvice.java
@@ -1,6 +1,6 @@
 package com.codesoom.assignment.advice;
 
-import com.codesoom.assignment.dto.ErrorResponse;
+import com.codesoom.assignment.advice.dto.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
