+++ app/src/main/java/com/codesoom/assignment/advice/dto/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.dto;
+package com.codesoom.assignment.advice.dto;
 
 public class ErrorResponse {
     private String message;
